refactor: routing new pages
