+++ src/components/App/Acciones/Form/index.jsx
@@ -1,0 +1,13 @@
+// components
+import { Card } from 'components/ui/Card'
+import { Heading } from '@chakra-ui/react'
+
+export const AccionesForm = () => {
+  return (
+    <>
+      <Card bg="red.100">
+        <Heading size="md"> Listado de Acciones </Heading>
+      </Card>
+    </>
+  )
+}
