+++ src/components/Auth/Calculator/Form/index.jsx
@@ -1,0 +1,13 @@
+// components
+import { Card } from 'components/ui/Card'
+import { Heading } from '@chakra-ui/react'
+
+export const CalculadoraForm = () => {
+  return (
+    <>
+      <Card bg="red.100">
+        <Heading size="md"> Calculator </Heading>
+      </Card>
+    </>
+  )
+}
