--- src/components/ui/Input/Email/index.jsx
@@ -1,19 +1,0 @@
-import { FormControl, FormLabel, Input, FormHelperText } from "@chakra-ui/react"
-
-export const InputEmail = ({ handleInput, info, error }) => {
-  return (
-    <>
-      <FormControl my={2}>
-        <FormLabel> Email </FormLabel>
-        <Input
-          name='email'
-          type='email'
-          placeholder='email'
-          onChange={handleInput}
-          value={info.email}
-        />
-         { error ? <FormHelperText color='red.400'> {error} </FormHelperText> : ''}
-      </FormControl>
-    </>
-  )
-}
