--- src/components/ui/Input/Password/Validator/index.jsx
@@ -1,14 +1,0 @@
-// components
-import { Text, Icon } from '@chakra-ui/react'
-// icons
-import { AiOutlineCheckCircle, AiFillCheckCircle } from 'react-icons/ai'
-
-export const PassValidator = ({ valid, value }) => {
-  return (
-    <>
-      <Text color={ !valid && 'gray.500'} fontSize="sm">
-        <Icon as={ !valid ? AiOutlineCheckCircle : AiFillCheckCircle} /> {value}
-      </Text>
-    </>
-  )
-}
