+++ src/pages/App/Acciones/index.jsx
@@ -1,0 +1,12 @@
+// components
+import { Heading } from '@chakra-ui/react'
+import { AccionesForm } from 'components/App/Acciones/Form'
+
+export const Acciones = () => {
+  return (
+    <>
+      <Heading> Acciones </Heading>
+      <AccionesForm/>
+    </>
+  )
+}
