--- src/pages/App/Calculator/index.jsx
@@ -1,12 +1,0 @@
-// components
-import { Heading } from '@chakra-ui/react'
-import { Form } from 'components/App/Calculator/Form-complete'
-
-export const Calculator = () => {
-  return (
-    <>
-      <Heading> Calculadora </Heading>
-      <Form />
-    </>
-  )
-}
