+++ src/pages/App/Daily/index.jsx
@@ -1,0 +1,12 @@
+// components
+import { Heading } from '@chakra-ui/react'
+import { Form } from 'components/App/Home/Form-complete'
+
+export const Daily = () => {
+  return (
+    <>
+      <Heading> Daily </Heading>
+      <Form />
+    </>
+  )
+}
