+++ src/pages/Auth/Calculator/index.jsx
@@ -2,7 +2,7 @@
 import { Heading, Button } from '@chakra-ui/react'
 import { Modal } from 'components/ui/Modal'
 import { RegisterTYP } from 'components/Auth/Register/TYP'
-import { Card } from 'components/ui/Card'
+import { CalculadoraForm } from 'components/Auth/Calculator/Form'
 // data
 import { Link } from 'react-router-dom'
 import { paths } from 'services/routes'
@@ -10,21 +10,19 @@ import { paths } from 'services/routes'
 import { useDisclosure } from '@chakra-ui/hooks'
 import { useEffect } from 'react'
 
-export const FirstCalculator = () => {
+export const Calculator = () => {
   // reminder
   const { isOpen: TYPiO, onOpen: TYPoO, onClose: TYPoC } = useDisclosure()
 
   useEffect(() => {
-    TYPoO()
+    // TYPoO()
   }, [TYPoO])
 
   return (
     <>
       <Heading> Bienvenido </Heading>
 
-      <Card bg='red.100'>
-        <Heading> Calculadora </Heading>
-      </Card>
+      <CalculadoraForm />
 
       <Link to={paths.home}>
         <Button size="lg"> Ver mi resultado </Button>
